backend/pyproject.toml
[tool.poetry.dependencies]
opentelemetry-api = "^1.27.0"
opentelemetry-sdk = "^1.27.0"
opentelemetry-exporter-otlp = "^1.27.0"
opentelemetry-instrumentation-fastapi = "^0.48b0"
opentelemetry-instrumentation-sqlalchemy = "^0.48b0"
opentelemetry-instrumentation-celery = "^0.48b0"
celery = "^5.3"
redis = "^5.0"
